pyproject.toml
[project]
name = "street-onion-pristine"
version = "0.1.0"
description = ""
authors = [
    {name = "vrZambrano",email = "victorzambrano13@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.25.1)",
    "selenium (>=4.0.0)",
    "boto3 (>=1.26.0)",
    "python-dotenv (>=0.19.0)",
    "pandas (>=1.3.0)",
    "pyarrow (>=5.0.0)",
    "openai (>=1.98.0,<2.0.0)"
]

[tool.poetry]
name = "street-onion-pristine"
version = "0.1.0"
description = ""
authors = ["vrZambrano <victorzambrano13@gmail.com>"]
packages = [
    { include = "src" },
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
